+++ src/guest/sh4/sh4.c
@@ -12,11 +12,11 @@
 #include "guest/rom/boot.h"
 #include "guest/rom/flash.h"
 #include "guest/scheduler.h"
+#include "imgui.h"
 #include "jit/frontend/sh4/sh4_fallback.h"
 #include "jit/frontend/sh4/sh4_frontend.h"
 #include "jit/frontend/sh4/sh4_guest.h"
 #include "jit/jit.h"
-#include "render/imgui.h"
 
 #if ARCH_X64
 #include "jit/backend/x64/x64_backend.h"
